# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mass Finder API",
        "description": "API para busca de paróquias e horários de missa",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{baseUrl}/api",
            "description": "API Base URL",
            "variables": {
                "baseUrl": {
                    "default": "http://localhost:8000"
                }
            }
        }
    ],
    "paths": {
...
        "/parish/profile": {
            "get": {
                "tags": [
                    "Parish"
                ],
                "summary": "Perfil da paróquia",
                "operationId": "fcb47e8f72f089bed0354f1dc17cef0f",
                "responses": {
                    "200": {
                        "description": "Perfil com horários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ParishDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paróquia não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Parish"
                ],
                "summary": "Atualizar perfil da paróquia",
                "operationId": "2a24546878a55ea7e3c7540512821cd9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateParishProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ParishDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paróquia não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ParishDetail": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Paróquia São José"
                    },
                    "cnpj": {
                        "type": "string",
                        "example": "12.345.678/0001-90"
                    },
                    "neighborhood": {
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "nullable": true
                    },
                    "mass_schedules": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MassSchedule"
                        }
                    }
                },
                "type": "object"
            },
            "MassSchedule": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "parish_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "day_of_week": {
                        "description": "0=Domingo, 1=Segunda, ..., 6=Sábado",
                        "type": "integer",
                        "maximum": 6,
                        "minimum": 0
                    },
                    "time": {
                        "type": "string",
                        "example": "08:00"
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateParishProfileRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Paróquia São José"
                    },
                    "neighborhood": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "GenericMessageResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Operação realizada com sucesso."
                    }
                },
                "type": "object"
            },
            "ErrorValidationResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtido via POST /api/auth/login",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}